requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dataclasses-json==0.6.4
distro==1.9.0
fastapi==0.110.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
httpx-sse==0.4.0
idna==3.6
isodate==0.6.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.9
langchain-cli==0.0.21
langchain-community==0.0.24
langchain-core==0.1.27
langserve==0.0.45
langsmith==0.1.10
lxml==5.1.0
markdown-it-py==3.0.0
marshmallow==3.21.0
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
pendulum==3.0.0
platformdirs==4.2.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
shellingham==1.5.4
simple-salesforce==1.12.5
six==1.16.0
slack-bolt==1.18.1
slack_sdk==3.27.1
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.27
sse-starlette==1.8.2
starlette==0.36.3
tavily-python==0.1.9
tenacity==8.2.3
time-machine==2.13.0
tomlkit==0.12.4
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.23.2
yarl==1.9.4
zeep==4.2.1
